+++ src/service/extensions/index.ts
@@ -1,3 +1,4 @@
+export * from './tatumsdk.wallet.providers.dto'
 export * from './tatumsdk.container'
 export * from './tatumsdk.extensions.dto'