+++ src/service/walletProvider/index.ts
@@ -1,1 +1,2 @@
 export * from './wallet.provider'
+export * from './metaMask'